y as a float

+++ diff_year.py
@@ -31,7 +31,7 @@ yearlist.set_index('id', inplace = True)
 yearlist = yearlist.sort_index()
 length = len(yearlist)
 
-y = ((abs(yearlist.loc[aid]-yearlist.loc[bid])/(maxyear-minyear))
+y = (float(abs(yearlist.loc[aid]-yearlist.loc[bid])/(maxyear-minyear))
     for (aid, bid) in it.combinations(idlist,2))
 
 x = np.array(list(y))
